Text.Builder

Documentation

data Builder

Specification of how to efficiently construct strict Text. Provides instances of Semigroup and Monoid, which have complexity of O(1).

Accessors

run :: Builder -> Text

Execute a builder producing a strict text

length :: Builder -> Int

Get the amount of characters

null :: Builder -> Bool

Check whether the builder is empty

Output IO

putToStdOut :: Builder -> IO ()

Put builder, to stdout

putToStdErr :: Builder -> IO ()

Put builder, to stderr

putLnToStdOut :: Builder -> IO ()

Put builder, followed by a line, to stdout

putLnToStdErr :: Builder -> IO ()

Put builder, followed by a line, to stderr

Constructors

Builder manipulators

intercalate :: Foldable foldable => Builder -> foldable Builder -> Builder

Intercalate builders

padFromLeft :: Int -> Char -> Builder -> Builder

Pad a builder from the left side to the specified length with the specified character

padFromRight :: Int -> Char -> Builder -> Builder

Pad a builder from the right side to the specified length with the specified character

Textual

text :: Text -> Builder

Strict text

lazyText :: Text -> Builder

Lazy text

string :: String -> Builder

String

asciiByteString :: ByteString -> Builder

ASCII byte string

hexData :: ByteString -> Builder

Hexadecimal readable representation of binary data.

Character

char :: Char -> Builder

Unicode character

Low-level character

unicodeCodePoint :: Int -> Builder

Unicode code point

utf16CodeUnits1 :: Word16 -> Builder

Single code-unit UTF-16 character

utf16CodeUnits2 :: Word16 -> Word16 -> Builder

Double code-unit UTF-16 character

utf8CodeUnits1 :: Word8 -> Builder

Single code-unit UTF-8 character

utf8CodeUnits2 :: Word8 -> Word8 -> Builder

Double code-unit UTF-8 character

utf8CodeUnits3 :: Word8 -> Word8 -> Word8 -> Builder

Triple code-unit UTF-8 character

utf8CodeUnits4 :: Word8 -> Word8 -> Word8 -> Word8 -> Builder

UTF-8 character out of 4 code units

Integers

Decimal

decimal :: Integral a => a -> Builder

Decimal representation of an integral value

unsignedDecimal :: Integral a => a -> Builder

Decimal representation of an unsigned integral value

thousandSeparatedDecimal :: Integral a => Char -> a -> Builder

Decimal representation of an integral value with thousands separated by the specified character

thousandSeparatedUnsignedDecimal :: Integral a => Char -> a -> Builder

Decimal representation of an unsigned integral value with thousands separated by the specified character

dataSizeInBytesInDecimal :: Integral a => Char -> a -> Builder

Data size in decimal notation over amount of bytes.

Binary

unsignedBinary :: Integral a => a -> Builder

Unsigned binary number

unsignedPaddedBinary :: (Integral a, FiniteBits a) => a -> Builder

Unsigned binary number

Hexadecimal

hexadecimal :: Integral a => a -> Builder

Hexadecimal representation of an integral value

unsignedHexadecimal :: Integral a => a -> Builder

Unsigned hexadecimal representation of an integral value

Digits

decimalDigit :: Integral a => a -> Builder

Decimal digit

hexadecimalDigit :: Integral a => a -> Builder

Hexadecimal digit

Real

fixedDouble

Arguments

:: Int
Amount of decimals after point.

-> Double

-> Builder

Double with a fixed number of decimal places.

doublePercent

Arguments

:: Int
Amount of decimals after point.

-> Double

-> Builder

Double multiplied by 100 with a fixed number of decimal places applied and followed by a percent-sign.

Time

intervalInSeconds :: RealFrac seconds => seconds -> Builder

Time interval in seconds. Directly applicable to DiffTime and NominalDiffTime.